refactor: Добавлен абстрактный класс для отображений
Все отображения в директории src/view/ теперь унаследованы от него.
